requirements.txt
streamlit
pymupdf
pytesseract
pillow
pandas
